Updated to include first acceptance test to prove out agouti framework, also include setup and tear down of suite and before/after test. Rename file to acceptance_test.
