--- test/acceptance/acceptance.go
@@ -1,34 +1,0 @@
-package acceptance
-
-import (
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-	"github.com/sclevine/agouti"
-
-	"testing"
-	"../common"
-	"time"
-)
-
-func TestAcceptance(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Acceptance Suite")
-}
-
-var agoutiDriver *agouti.WebDriver
-
-var _ = BeforeSuite(func() {
-	// Choose a WebDriver:
-
-	agoutiDriver = agouti.PhantomJS()
-
-	go common.StartPageCompositionServer(":5001")
-	go common.StartStubServer(":5002")
-
-	time.Sleep(120 * time.Second)
-	Expect(agoutiDriver.Start()).To(Succeed())
-})
-
-var _ = AfterSuite(func() {
-	Expect(agoutiDriver.Stop()).To(Succeed())
-})
